refactor(giselle-node): move node actions to dedicated module
- Move node action types and reducer from graph/v2/node to giselle-node/actions
- Update imports across codebase to use new location
- Remove duplicate node action definitions
- Clean up file organization for better domain separation
- Delete deprecated graph/v2/node.ts file
